# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Treffin API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /feed:
    get:
      operationId: getFeed
      tags: [feed]
      summary: Get home feed posts
      parameters:
        - name: tab
          in: query
          schema:
            type: string
            enum: [for_you, following, debates, articles, communities]
          required: false
        - name: page
          in: query
          schema:
            type: integer
          required: false
        - name: authorId
          in: query
          schema:
            type: integer
          required: false
      responses:
        "200":
          description: Feed posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/FeedPost"
...
  /posts:
    post:
      operationId: createPost
      tags: [feed]
      summary: Create a new post/thought
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostInput"
      responses:
        "201":
          description: Created post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPost"
...
  /posts/{id}/reveal:
    patch:
      operationId: revealPost
      tags: [feed]
      summary: Reveal the author identity of an anonymous post (owner only)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post with identity revealed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPost"
        "403":
          description: Not the post owner
        "404":
          description: Post not found
  /posts/{id}/like:
    post:
      operationId: likePost
      tags: [feed]
      summary: Like or unlike a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FeedPost"
  /posts/{id}/comments:
    get:
      operationId: getPostComments
      tags: [feed]
      summary: Get comments for a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Post comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DebateComment"
        "404":
          description: Post not found
    post:
      operationId: createPostComment
      tags: [feed]
      summary: Post a comment on a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DebateCommentInput"
      responses:
        "201":
          description: Created comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DebateComment"
        "404":
          description: Post not found
  /posts/{id}/report:
    post:
      operationId: reportPost
      tags: [feed]
      summary: Report a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PostReportInput"
      responses:
        "200":
          description: Report recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOk"
        "404":
          description: Post not found
  /posts/{id}/comments/{commentId}/like:
    post:
      operationId: likePostComment
      tags: [feed]
      summary: Toggle like on a community post comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated like state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommentLikeToggle"
        "401":
          description: Not authenticated
        "404":
          description: Comment not found
  /posts/{id}/comments/{commentId}:
    delete:
      operationId: deletePostComment
      tags: [feed]
      summary: Delete a comment on a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
        - name: commentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminOk"
        "404":
          description: Comment not found
...
components:
  schemas:
...
    FeedPost:
      type: object
      required: [id, type, authorId, authorName, authorTitle, authorAvatar, createdAt, likes, comments, reposts]
      properties:
        id:
          type: integer
        type:
          type: string
          enum: [opinion, article, knowledge, debate_highlight, quote]
        authorId:
          type: integer
        authorName:
          type: string
        authorTitle:
          type: string
        authorAvatar:
          type: ["string", "null"]
        isVerified:
          type: boolean
        createdAt:
          type: string
        content:
          type: ["string", "null"]
        title:
          type: ["string", "null"]
        excerpt:
          type: ["string", "null"]
        imageUrl:
          type: ["string", "null"]
        topic:
          type: ["string", "null"]
        readTime:
          type: ["integer", "null"]
        likes:
          type: integer
        comments:
          type: integer
        reposts:
          type: integer
        saved:
          type: boolean
        isAnonymous:
          type: boolean
        isOwner:
          type: boolean
        liked:
          type: boolean
...
    PostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        type:
          type: string
          enum: [opinion, article, debate, poll]
        topic:
          type: string
        isAnonymous:
          type: boolean
...
    DebateComment:
      type: object
      required: [id, authorId, authorName, content, createdAt, isRemoved]
      properties:
        id:
          type: integer
        authorId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        side:
          type: ["string", "null"]
        isFlagged:
          type: boolean
        flagLabel:
          type: ["string", "null"]
        createdAt:
          type: string
        editedAt:
          type: ["string", "null"]
        isRemoved:
          type: boolean
        removedReason:
          type: ["string", "null"]
        sources:
          type: ["string", "null"]
          description: JSON array of {url, label} source citations
        wordCount:
          type: ["integer", "null"]
        personalAttackWarning:
          type: ["string", "null"]
          description: Non-blocking advisory returned when post contains second-person attack patterns
        likes:
          type: integer
          description: Total like count for this comment
        likedByMe:
          type: boolean
          description: Whether the authenticated user has liked this comment
        parentCommentId:
          type: ["integer", "null"]
          description: ID of the parent comment if this is a reply
    CommentLikeToggle:
      type: object
      required: [likes, liked]
      properties:
        likes:
          type: integer
        liked:
          type: boolean
    AdminOk:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
...
    DebateCommentInput:
      type: object
      required: [authorName, content]
      properties:
        authorId:
          type: integer
        authorName:
          type: string
        content:
          type: string
        side:
          type: string
        sources:
          type: string
          description: JSON array of {url, label} source citations
        parentCommentId:
          type: integer
          description: ID of parent comment when posting a reply
...
    PostReportInput:
      type: object
      properties:
        reason:
          type: string
...